+++ src/resources/src/js/components/partials/Footer.svelte
@@ -13,7 +13,7 @@
                 class="btn disabled submit"
                 aria-disabled={$isBusy || !$insertion}
                 class:disabled={$isBusy || !$insertion}
-                on:click={insert}>
+                on:click={sendToField('insert')}>
                 <div class="label">Insert Results</div>
                 <div class="spinner spinner-absolute"></div>
             </button>
@@ -33,22 +33,24 @@
             class:is-active={dropdownActive}>
             <li>
                 <button type="button"
-                    on:click|stopPropagation={prepend}>
+                    on:click|stopPropagation={sendToField('prepend')}>
                     Prepend Results
                 </button>
             </li>
             <li>
                 <button type="button"
-                    on:click|stopPropagation={append}>
+                    on:click|stopPropagation={sendToField('append')}>
                     Append Results
                 </button>
             </li>
+            <!-- {#if $field.redactor} -->
             <li>
                 <button type="button"
-                    on:click|stopPropagation={replace}>
+                    on:click|stopPropagation={sendToField('replace')}>
                     Replace with Results
                 </button>
             </li>
+            <!-- {/if} -->
             <li>
                 <button type="button"
                     on:click|stopPropagation={clipboard}>
@@ -60,64 +62,14 @@
 </div>
 
 <script context="module">
-    import { getContext } from 'svelte';
-    import snarkdown from '@bpmn-io/snarkdown';
-    import { isActive, isBusy, insertion } from '../../store';
+    import { field, isActive, isBusy, insertion } from '../../store';
 </script>
 
 <script>
     export let dropdownActive;
 
-    const redactor = getContext('redactor');
-
-    function prepare(content) {
-        content = redactor.cleaner.paragraphize(content);
-        content = snarkdown(content);
-        content = redactor.cleaner.input(content);
-
-        return content;
-    }
-
-    function insert() {
-        if (!$insertion) {
-            return;
-        }
-
-        redactor.insertion.insertHtml(prepare($insertion));
-        $isActive = false;
-    }
-
-    function append() {
-        if (!$insertion) {
-            return;
-        }
-
-        redactor.insertion.set([
-            redactor.api('source.getCode'),
-            prepare($insertion)
-        ].join(`\n`));
-        $isActive = false;
-    }
-
-    function prepend() {
-        if (!$insertion) {
-            return;
-        }
-
-        redactor.insertion.set([
-            prepare($insertion),
-            redactor.api('source.getCode')
-        ].join(`\n`));
-        $isActive = false;
-    }
-
-    function replace() {
-        if (!$insertion) {
-            return;
-        }
-
-        redactor.insertion.set(prepare($insertion));
-        $isActive = false;
+    function sendToField(method) {
+        return () => $isActive = !$field[method]($insertion);
     }
 
     function clipboard() {
@@ -128,12 +80,13 @@
         document.addEventListener('copy', copy);
         document.execCommand('copy');
         document.removeEventListener('copy', copy);
+
         $isActive = false;
     }
 
     function copy(event) {
-        event.clipboardData.setData('text/html', prepare($insertion));
-        event.clipboardData.setData('text/plain', prepare($insertion));
+        event.clipboardData.setData('text/html', field._prepare($insertion));
+        event.clipboardData.setData('text/plain', field._prepare($insertion));
         event.preventDefault();
     }
 </script>