+++ src/resources/src/js/store.js
@@ -1,6 +1,9 @@
 import throttle from './modules/throttle';
 import { writable, readable } from 'svelte/store';
 
+export const enabledFields = writable([]);
+export const field = writable(null);
+export const preview = writable('');
 export const errors = writable([]);
 export const isBusy = writable(false);
 export const isActive = writable(false);